Move lag into own file
