+++ irksome/__init__.py
@@ -14,7 +14,8 @@ from .tableaux.multistep_tableaux import (
     AdamsMoulton,
 )
 from .tableaux.pep_explicit_rk import PEPRK
-from .ufl.deriv import Dt, expand_time_derivatives, lag, check_irksome_import_order
+from .ufl.deriv import Dt, expand_time_derivatives, check_irksome_import_order
+from .ufl.lag import lag
 
 check_irksome_import_order()
