+++ irksome/tools.py
@@ -6,7 +6,8 @@ from ufl import as_tensor, replace as ufl_replace
 
 import FIAT
 
-from .ufl.deriv import TimeDerivative, lag_label
+from .ufl.deriv import TimeDerivative
+from .ufl.lag import lag_label
 
 
 def dot(A, B):
@@ -122,7 +123,7 @@ def getNullspace(V, Vbig, num_stages, nullspace):
 
 def replace(e, mapping):
     """A wrapper for ufl.replace that allows numpy arrays and skips
-    substitution into sub-expressions wrapped by :func:`~.lag`."""
+    substitution into sub-expressions wrapped by :func:`~irksome.lag`."""
     cmapping = {k: as_tensor(v) for k, v in mapping.items()}
     for var in extract_type(e, Variable):
         if var.ufl_operands[1] is lag_label:
